# document load time: CC Libraries optimization