--- a/src/js/actions/libraries.js
+++ b/src/js/actions/libraries.js
@@ -55,13 +55,6 @@ define(function (require, exports) {
         historyActions = require("./history"),
         preferencesActions = require("./preferences"),
         LibrarySyncStatus = require("js/models/library_sync_status");
-
-    /**
-     * The external CC Libraries API is loaded here asynchronously, used in beforeStartup
-     *
-     * @type {Promise}
-     */
-    var apiLoadPromise = Promise.promisify($S)("file://shared/libs/cc-libraries-api.min.js");
         
     /**
      * For image elements, their extensions signify their representation type
@@ -1273,14 +1266,35 @@ define(function (require, exports) {
         _toolModalStateChangedHandler = function (event) {
             var isPlacingGraphic = this.flux.store("library").getState().isPlacingGraphic,
                 modalStateEnded = event.state && event.state._value === "exit";
-
+        
             if (isPlacingGraphic && modalStateEnded) {
                 this.flux.actions.libraries.handleCompletePlacingGraphic();
             }
         }.bind(this);
         descriptor.addListener("toolModalStateChanged", _toolModalStateChangedHandler);
 
-        return apiLoadPromise
+        return Promise.resolve();
+    };
+    beforeStartup.action = {
+        reads: [locks.JS_PREF],
+        writes: [locks.JS_LIBRARIES, locks.CC_LIBRARIES]
+    };
+
+    /**
+     * Libraries loadedCollections event handlers.
+     *
+     * @private
+     * @type {function}
+     */
+    var _handleLibrariesLoadedHelper;
+
+    /**
+     * After startup, load the libraries
+     *
+     * @return {Promise}
+     */
+    var afterStartup = function () {
+        return Promise.promisify($S)("file://shared/libs/cc-libraries-api.min.js")
             .timeout(3000, "CC Libraries API load timeout, please don't restart and notify the chatroom!")
             .bind(this)
             .then(function () {
@@ -1288,14 +1302,14 @@ define(function (require, exports) {
                 // There is now a ccLibraries object in this scope
                 // So we emit that to the store
                 this.dispatch(events.libraries.LIBRARIES_API_LOADED, ccLibraries);
-
+        
                 var dependencies = {
                     // Photoshop on startup will grab the port of the CC Library process and expose it to us
                     vulcanCall: function (requestType, requestPayload, responseType, callback) {
                         descriptor.getProperty("application", "designSpaceLibrariesInfo")
                             .then(function (imsInfo) {
                                 var port = imsInfo.port;
-
+        
                                 callback(JSON.stringify({ port: port }));
                             });
                     },
@@ -1331,7 +1345,7 @@ define(function (require, exports) {
                         }
                     }
                 };
-
+        
                 // SHARED_LOCAL_STORAGE flag forces websocket use
                 ccLibraries.configure(dependencies, {
                     SHARED_LOCAL_STORAGE: true,
@@ -1344,49 +1358,31 @@ define(function (require, exports) {
                         ELEMENT_COLORTHEME_TYPE
                     ]
                 });
+
+                if (_handleLibrariesLoadedHelper) {
+                    ccLibraries.removeLoadedCollectionsListener(_handleLibrariesLoadedHelper);
+                }
+
+                // Listen to the load event of CC Libraries. The event has two scenarios:
+                //     loaded: Libraries data is ready for use. Fired after user sign in creative cloud.
+                //     unloaded: Libraries data is cleared. Fired after user sign out creative cloud.
+                _handleLibrariesLoadedHelper = function () {
+                    return this.flux.actions.libraries.handleLibrariesLoaded();
+                }.bind(this);
+                ccLibraries.addLoadedCollectionsListener(_handleLibrariesLoadedHelper);
+                
+                // ccLibraries will not notify if users is logged out before, call the callback manually instead.
+                setTimeout(function () {
+                    if (ccLibraries.getLoadedCollections.length === 0) {
+                        _handleLibrariesLoadedHelper();
+                    }
+                }, 1000);
             });
-    };
-    beforeStartup.action = {
-        reads: [locks.JS_PREF],
-        writes: [locks.JS_LIBRARIES, locks.CC_LIBRARIES]
-    };
-
-    /**
-     * Libraries loadedCollections event handlers.
-     *
-     * @private
-     * @type {function}
-     */
-    var _handleLibrariesLoadedHelper;
-
-    /**
-     * After startup, load the libraries
-     *
-     * @return {Promise}
-     */
-    var afterStartup = function () {
-        var ccLibraries = this.flux.store("library").getLibrariesAPI();
-
-        if (_handleLibrariesLoadedHelper) {
-            ccLibraries.removeLoadedCollectionsListener(_handleLibrariesLoadedHelper);
-        }
-
-        // Listen to the load event of CC Libraries. The event has two scenarios:
-        //     loaded: Libraries data is ready for use. Fired after user sign in creative cloud.
-        //     unloaded: Libraries data is cleared. Fired after user sign out creative cloud.
-        _handleLibrariesLoadedHelper = function () {
-            return this.flux.actions.libraries.handleLibrariesLoaded();
-        }.bind(this);
-        ccLibraries.addLoadedCollectionsListener(_handleLibrariesLoadedHelper);
-        
-        // Trigger the load event callback manually for initial start up.
-        return this.transfer(handleLibrariesLoaded);
     };
     afterStartup.action = {
         reads: [locks.JS_LIBRARIES, locks.CC_LIBRARIES],
         writes: [],
-        modal: true,
-        transfers: [handleLibrariesLoaded]
+        modal: true
     };
     
     /**