--- a/src/style/shared/icons.less
+++ b/src/style/shared/icons.less
@@ -210,7 +210,7 @@ svg use {
 }
 
 .layer__not-visible
-.layer__not-invisible + .layer-group {
+.layer__not-visible + .layer-group {
     .face:hover .button-toggle,
     .face .button-toggle {
         svg {