--- a/src/style/shared/link-list.less
+++ b/src/style/shared/link-list.less
@@ -20,12 +20,6 @@
  * DEALINGS IN THE SOFTWARE.
  *
  */
-.null-state{
-    .panel:hover .panel__hide{
-        display: none;
-    }
-}
-
 .link-list__list {
 	width: 100%;
 }