--- a/src/style/shared/loader.less
+++ b/src/style/shared/loader.less
@@ -21,7 +21,7 @@
  *
  */
 
-.loader-animation svg use{    
+.loader-animation svg g{
     animation: loader-spin;
     animation-duration: 0.66s; 
     animation-timing-function: steps(6,end);