--- a/src/style/shared/title-header.less
+++ b/src/style/shared/title-header.less
@@ -94,7 +94,7 @@
     line-height: 2rem;
 }
 
-.panel-set__container__small-screen {
+.panel-set__small-screen {
     .appearance,
     .effects,
     .export,