… src/js/jsx/PanelSet.jsx