This PR includes the following optimizations on CC Libraries to improve performance.

* Defer CC Libraries loading/rendering process to improve startup performance. Before, CC Libraries is loaded in the `beforeStartup` callback, which slows down the startup process and triggers libraries panel rendering, which slow down the startup further. This process can be deferred as the chance of using the librarise panel right after startup is small. 
* Prevent the Libraries panel DOM from removing when closing the last document. Otherwise, we have to waste time in rendering the panel again when creating a new document.
* Pre-render the Libraries panel in no-doc state. While in no-doc state, the libraries panel can be pre-rendered (by React) when the data is ready. 

Below is the result of the before/after test in various scenarios:

|  | Before | After | Diff |
| --- | --- | --- | --- |
| no-doc startup | 869ms | **758ms** | -111ms |
| empty doc startup | 1741ms | **828ms** | -903ms |
| vermillion startup | 2562ms | **1640ms** | -922ms |
| no-doc -> empty doc (scripting) | 691ms | **252ms** | -439ms |
| no-doc -> empty doc (rendering) | 71ms | **54ms** | -17ms |
| empty doc -> no-doc (scripting) | 208ms | 208ms | 0 |
| empty doc -> no-doc (rendering) | **5ms**  | 9ms | 4ms |
(selected library: `Samples Collection April 23 2015`)