@iwehrman 2016-02-04
One concern with this is that the rest of the panel contents fade in when they're ready, but the Libraries panel now does not, and it feels just a little off. Can you do some experiments to try to improve this? Like, try fading in the libraries panel content when it's ready?