@iwehrman 2016-02-04
Or, alternatively, rendering a spinner when first showing the UI and replacing that spinner with the contents when ready?