@shaoshing 2016-02-04
Good idea. I will experiment with the spinner loader if content is not ready