@iwehrman 2016-02-04
Sorry, I started reading through this and then got distracted, but besides my one gripe I think this looks great. I'm ready to merge if we can improve the loading experience just a little bit.