@shaoshing 2016-02-05
@iwehrman I experimented and compared the `fade-in` and `spinner` solutions, and I think `face-in` has better user experience in this case. The libraries panel usually can finish loading and rendering within within 500ms (even for large collection, like `Samples Collection April 23 2015`, the time is still within 1200ms range). In such short amount of time, the `spinner` seems to be distracting. Also, it makes me feel that the panel is slower, compare to fading-in the content when ready. The reason might be because moving object can attract attension easier.

I've push my changes with the fade-in solution, please try it out and let me know what you feel.