@iwehrman 2016-02-07
I'm not surprised that the spinner was distracting, but I'm glad you tried both options. I think the current behavior is great. I'm ready to merge except for that one additional typo. :+1: :+1: