@shaoshing 2016-02-09
@iwehrman I just addressed the typo. Please review again.