@shaoshing 2016-02-09
Thanks for merging!